More rework of testSysbox* script for idmapped mount testing.
These chagnes are needed due to a change in sysbox-runc that
enables it to use shiftfs in combination with idmapped mounts
in hosts that support both mechanism (i.e., shiftfs is used
for the container's rootfs, idmapped mounts for other mounts).

This change also refactors code in the testSysbox* scripts
to make them easier to understand and maintain.

Signed-off-by: Cesar Talledo <ctalledo@nestybox.com>